AVE/requirements.txt
certifi==2022.9.24
deepspeed==0.7.3
einops==0.4.1
gpuinfo==1.0.0a7
h5py==2.10.0
ipdb==0.13.9
loralib==0.1.1
matplotlib==3.6.2
moviepy==1.0.3
munch==2.5.0
numpy==1.23.1
pandas==1.5.0
Pillow==9.3.0
pretrainedmodels==0.7.4
scikit_learn==1.2.0
scipy==1.9.1
SoundFile==0.10.3.post1
timm==0.6.7
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
tqdm==4.40.0
transformers==4.24.0
wandb==0.13.3
wget==3.2
